moved HttpGetRootCall to commons
